+++ app/javascript/components/GameListSection.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
-import { CaretDownFill, CaretUp, CaretUpFill } from "react-bootstrap-icons";
 import { getAPI } from "../utilities/network";
 import GameListRow, { GameRowData } from "./GameListRow";
+import { NavDown, NavUp } from "./utilities/buttons";
 
 interface GameListSectionProps {
   scope: string;
@@ -68,17 +68,25 @@ export default function GameListSection({ scope, user, queue, shiftQueue }: Game
   }, [queue, queue.length])
 
   const scrollUp = scroll.up ?
-    <div onClick={() => setPage(page - 1)}><CaretUpFill /></div> :
-    (scroll.down ? <div><CaretUp /></div> :
-      <div className="transparent"><CaretUpFill /></div>)
+    <div className="main-page-list-nav" onClick={() => setPage(page - 1)}>{ NavUp(true) }</div> :
+      (scroll.down ? <div className="main-page-list-nav">{ NavUp(false) }</div> : "")
 
   const scrollDown = scroll.down ?
-    <div onClick={() => setPage(page + 1)}><CaretDownFill /></div> : ""
+    <div className="main-page-list-nav" onClick={() => setPage(page + 1)}>{ NavDown(true) }</div> : ""
+
+  const scrollMiddle = scroll.up || scroll.down ?
+    <div className="flex-fill main-page-list-scroll">
+      <div className="main-page-list-not-scroll-inside"></div>
+    </div> : <div className="flex-fill main-page-list-scroll-filler"></div>
 
   return (
     <div className="flex">
       <div className="flex-fill">{ games.map((g, i) => <GameListRow key={i} data={g} />) }</div>
-      <div className="ml05em control-large">{scrollUp}{scrollDown}</div>
+      <div className="ml05em control-large flex-vertical">
+        {scrollUp}
+        {scrollMiddle}
+        {scrollDown}
+      </div>
     </div>
   )
 }